\name{Boi-package}
\alias{Boi-package}
\alias{Boi}
\docType{package}
\title{
\packageTitle{Boi}
}
\description{
\packageDescription{Boi}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{Boi}

Maintainer: \packageMaintainer{Boi}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = Boi)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{Boi}}
%\section{Documentation Index}{\packageIndices{Boi}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
